VncClientProtocol: use 32 bit for message sizes
Messages will never exceed a size of 4 GB and QByteArray::size() returns
int anyway so we can safely use 32 bit for size variables and arguments
everywhere.

+++ core/include/VncClientProtocol.h
@@ -109,7 +109,7 @@ private:
 	bool receiveResizeFramebufferMessage();
 	bool receiveXvpMessage();
 
-	bool readMessage( qint64 size );
+	bool readMessage( int size );
 
 	bool handleRect( QBuffer& buffer, rfbFramebufferUpdateRectHeader rectHeader );
 	bool handleRectEncodingRRE( QBuffer& buffer, uint bytesPerPixel );
